% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageIdentifier.R
\name{identifyLanguage}
\alias{identifyLanguage}
\title{Identifies the language of a given text based on frequent trigrams}
\usage{
identifyLanguage(text)
}
\arguments{
\item{text}{String for which the language should be identified}
}
\value{
Name of the detected language.
}
\description{
This function identifies the language of a given string based on the most
frequent trigrams in different languages. Supported languages are Czech,
Dutch, English, French, German, Spanish, Latvian and Lithuanian.
}
\examples{

text <- "The tea in the cup still is hot."
language <- identifyLanguage(text)
language

}
