% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderanalyzer-package.R
\name{orderanalyzer-package}
\alias{orderanalyzer-package}
\alias{orderanalyzer}
\title{Extracting order position tables from PDF-based order documents}
\description{
This packages provides functions for extracting text and order-position-tables from
PDF-based order documents.
}
\details{
\tabular{ll}{ Package: \tab orderanalyzer\cr Type: \tab Package\cr Version:
\tab 1.0.0\cr Date: \tab 2024-12-11\cr License: \tab GPL-3\cr Depends: \tab
R (>= 4.3.0)}
}
\author{
Michael Scholz \email{michael.scholz@th-deg.de}

Joerg Bauer \email{joerg.bauer@th-deg.de}
}
